Rename copy question file to copy content
