+++ apps/prairielearn/src/pages/instructorQuestionSettings/instructorQuestionSettings.ts
@@ -20,7 +20,7 @@ import {
 
 import { b64EncodeUnicode } from '../../lib/base64-util.js';
 import { config } from '../../lib/config.js';
-import { copyQuestionBetweenCourses } from '../../lib/copy-question.js';
+import { copyQuestionBetweenCourses } from '../../lib/copy-content.js';
 import { EnumGradingMethodSchema } from '../../lib/db-types.js';
 import {
   FileModifyEditor,
