+++ apps/prairielearn/src/pages/publicQuestionPreview/publicQuestionPreview.ts
@@ -4,7 +4,7 @@ import { z } from 'zod';
 
 import * as error from '@prairielearn/error';
 
-import { setQuestionCopyTargets } from '../../lib/copy-question.js';
+import { setQuestionCopyTargets } from '../../lib/copy-content.js';
 import { IdSchema, UserSchema } from '../../lib/db-types.js';
 import { features } from '../../lib/features/index.js';
 import {
